Fix linting and tests move to test folder
