+++ .eslintrc.json
@@ -14,11 +14,6 @@
         "ecmaVersion": 2018,
         "sourceType": "module"
     },
-    "extends": ["airbnb-base"],
     "rules": {
-        "no-shadow": "off",
-        "no-param-reassign": "off",
-        "eol-last": "off",
-        "arrow-parens": "off"
-      }
+    }
 }
